Add `TargetPort` to `RouteToApp` & use it to route connections to multi-port TCP apps
* Add TargetPort to RouteToApp and AppMetadata proto messages

* Pass TargetPort during cert generation

* Refactor Pack.makeTLSConfig to accept struct

This will make it easier to add targetPort to it.

* Add labels to UUIDs used by appaccess test pack app servers

This makes them easier to distinguish when routing doesn't work as expected.

* Refactor Pack.CreateAppSession to accept a struct

* TestTCP: Create app session within test

If we kept the old code, we'd need to manually create a session for each
target port, which would create a lot of duplication.

* Prepare integration test fixtures for multi-port tests

* Add api/utils/net.IsPortInRange

* Use TargetPort when routing TCP connections

* Inline dialMultiPortTCPApp, centralize logic

* Check target port when connecting to single-port app

* Reorder check in IsPortInRange

* Use int instead of uint16

* Extract picking dialTarget to separate function

* Improve err msg for single-port apps when targetPort != uriPort

* Fix unnecessary conversion to int

+++ lib/config/fileconf.go
@@ -2092,11 +2092,12 @@ type AppAWS struct {
 // PortRange can be used to describe a single port in which case the Port field is the port and the
 // EndPort field is 0.
 type PortRange struct {
-	// Port describes the start of the range. It must be greater than 0.
-	Port uint16 `yaml:"port"`
+	// Port describes the start of the range. It must be between 1 and 65535.
+	Port int `yaml:"port"`
 	// EndPort describes the end of the range, inclusive. When describing a port range, it must be
-	// greater than 2 and be greater than Port. When describing a single port, it must be set to 0.
-	EndPort uint16 `yaml:"end_port,omitempty"`
+	// greater than Port and less than or equal to 65535. When describing a single port, it must be
+	// set to 0.
+	EndPort int `yaml:"end_port,omitempty"`
 }
 
 // Proxy is a `proxy_service` section of the config file:
